requirements.txt
cheroot==8.5.2
click==7.1.2
colorama==0.4.4
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==20.0.4
idna==3.1
itsdangerous==1.1.0
jaraco.functools==3.2.1
Jinja2==2.11.3
MarkupSafe==1.1.1
more-itertools==8.7.0
numpy==1.20.1
Pillow==8.1.2
psycopg2-binary==2.8.6
qrcode==6.1
six==1.15.0
SQLAlchemy==1.4.1
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
